Move small-screen camera logic into gfx_camera.cpp
